refactor: move createOnSearchResults function to a dedicated search module
